+++ lib/trello_cli/cli/card/create.rb
@@ -10,13 +10,12 @@ module TrelloCli
         def run
           option_parser.parse!
 
-          card        = create_card
-          name        = card.attributes[:name]
-          description = card.attributes[:description]
+          card = create_card
+          data = { id:   card.attributes[:id],
+                   name: card.attributes[:name],
+                   desc: card.attributes[:desc] }
 
-          puts "Card Created."
-          puts "Name        : #{name}"
-          puts "Description : #{description}"
+          puts TrelloCli::Formatters::CardCreate.new(data).output @options[:output]
         end
 
         private
@@ -35,7 +34,7 @@ module TrelloCli
             end
 
             opts.on("-d", "--description [DESCRIPTION]", "Description Of Card") do |d|
-              @options[:description] = d
+              @options[:desc] = d
             end
 
             opts.on("-l", "--list [LIST]", "List Of Card") do |l|
@@ -48,6 +47,10 @@ module TrelloCli
 
             opts.on("-n", "--name [NAME]", "Name Of Card") do |n|
               @options[:name] = n
+            end
+
+            opts.on("-o", "--output [OUTPUT]", "Output format [json|tsv|legacy]." ) do |o|
+              @options[:output] = o
             end
           end
         end