+++ lib/trello_cli/cli/card/list.rb
@@ -10,11 +10,11 @@ module TrelloCli
         def run
           option_parser.parse!
 
-          list_cards.each do |card|
-            puts "| Name: #{card.name} ( #{card.id} )"
-            puts "| Description: #{card.description}" if @options[:description]
-            puts "|------------------------"
+          data = list_cards.map do |c|
+            { name: c.name, id: c.id, desc: c.desc }
           end
+
+          puts TrelloCli::Formatters::CardList.new(data).output(@options[:output])
         end
 
         private
@@ -35,8 +35,8 @@ module TrelloCli
               @options[:list_id] = l
             end
 
-            opts.on("-d", "--description", "Include description." ) do |o|
-              @options[:description] = o
+            opts.on("-o", "--output [OUTPUT]", "Output format [json|tsv|legacy]." ) do |o|
+              @options[:output] = o
             end
           end
         end