+++ lib/trello_cli/formatters/base.rb
@@ -1,0 +1,30 @@
+require 'json'
+
+module TrelloCli
+  module Formatters
+    class Base
+      def initialize(data)
+        @data = data
+      end
+
+      def output(format)
+        case format
+        when "json"
+          self.to_json
+        when "tsv"
+          self.to_tsv
+        else
+          self.to_legacy
+        end
+      end
+
+      def to_json
+        data.to_json
+      end
+
+      private
+
+      attr_reader :data
+    end
+  end
+end